refactoring
move handled exceptions to utils package
