+++ bzt/cli.py
@@ -31,11 +31,11 @@ from urllib.error import HTTPError
 import yaml
 from colorlog import ColoredFormatter
 
-from bzt import ManualShutdown, NormalShutdown, RCProvider, TaurusException, AutomatedShutdown
-from bzt import TaurusInternalException, TaurusConfigError, TaurusNetworkError, ToolError
+from bzt import ManualShutdown, NormalShutdown, RCProvider, AutomatedShutdown
 from bzt.engine import Engine, Configuration, SETTINGS, EXEC
 from bzt.linter import ConfigurationLinter
 from bzt.utils import get_stacktrace, is_int, BetterDict, is_url, RESOURCES_DIR, VERSION, GIT_INFO, BUILD_NUM
+from bzt.utils import TaurusException, TaurusInternalException, TaurusConfigError, TaurusNetworkError, ToolError
 
 
 class CLI(object):
@@ -616,7 +616,7 @@ class OptionParserWithAliases(OptionParser, object):
 
 def get_option_parser():
     usage = "Usage: bzt [options] [configs] [-aliases]"
-    dsc = "BlazeMeter Taurus Tool v%s, the configuration-driven test running engine" % bzt.VERSION
+    dsc = "BlazeMeter Taurus Tool v%s, the configuration-driven test running engine" % VERSION
     parser = OptionParserWithAliases(usage=usage, description=dsc, prog="bzt")
     parser.add_option('-l', '--log', action='store', default=None,
                       help="Log file location")
