+++ bzt/utils/build.py
@@ -1,0 +1,1 @@
+BUILD_NUM = "n/a"
