+++ bzt/utils/gitinfo.py
@@ -1,0 +1,1 @@
+GIT_INFO = "n/a"
