+++ bzt/utils/release.py
@@ -1,0 +1,1 @@
+VERSION = "1.15.0"
